+++ how-to/workspace-platform-starter/client/src/framework/shapes/setting-shapes.ts
@@ -21,6 +21,7 @@ import type { PlatformProviderOptions } from "./platform-shapes";
 import type { SplashScreenProviderOptions } from "./splash-shapes";
 import type { StorefrontProviderOptions } from "./store-shapes";
 import type { ThemeProviderOptions } from "./theme-shapes";
+import type { TrayProviderOptions } from "./tray-shapes";
 import type { VersionProviderOptions } from "./version-shapes";
 
 /**
@@ -151,4 +152,9 @@ export interface CustomSettings {
 	 * The settings you need to set if you want to indicate this platform supports favorites
 	 */
 	favoriteProvider?: FavoriteProviderOptions;
+
+	/**
+	 * The settings you need to set if you want to include a tray icon with menu
+	 */
+	trayProvider?: TrayProviderOptions;
 }